move database checker to another package
